context-engineering-intro/requirements.txt
anthropic>=0.40.0
python-dotenv>=1.0.0
tiktoken>=0.7.0
